Restructure action folder structure
