+++ src/main/kotlin/dev/mizarc/waystonewarps/WaystoneWarps.kt
@@ -3,14 +3,14 @@ package dev.mizarc.waystonewarps
 import co.aikar.commands.PaperCommandManager
 import dev.mizarc.waystonewarps.application.actions.teleport.LogPlayerMovement
 import dev.mizarc.waystonewarps.application.actions.teleport.TeleportPlayer
-import dev.mizarc.waystonewarps.application.actions.warp.BreakWarpBlock
-import dev.mizarc.waystonewarps.application.actions.warp.CreateWarp
-import dev.mizarc.waystonewarps.application.actions.warp.GetPlayerWarpAccess
-import dev.mizarc.waystonewarps.application.actions.warp.GetWarpAtPosition
-import dev.mizarc.waystonewarps.application.actions.warp.GetWarpPlayerAccess
-import dev.mizarc.waystonewarps.application.actions.warp.RefreshAllStructures
-import dev.mizarc.waystonewarps.application.actions.warp.UpdateWarpIcon
-import dev.mizarc.waystonewarps.application.actions.warp.UpdateWarpName
+import dev.mizarc.waystonewarps.application.actions.world.BreakWarpBlock
+import dev.mizarc.waystonewarps.application.actions.world.CreateWarp
+import dev.mizarc.waystonewarps.application.actions.discovery.GetPlayerWarpAccess
+import dev.mizarc.waystonewarps.application.actions.world.GetWarpAtPosition
+import dev.mizarc.waystonewarps.application.actions.discovery.GetWarpPlayerAccess
+import dev.mizarc.waystonewarps.application.actions.world.RefreshAllStructures
+import dev.mizarc.waystonewarps.application.actions.management.UpdateWarpIcon
+import dev.mizarc.waystonewarps.application.actions.management.UpdateWarpName
 import dev.mizarc.waystonewarps.application.services.*
 import dev.mizarc.waystonewarps.domain.discoveries.DiscoveryRepository
 import dev.mizarc.waystonewarps.domain.playerstate.PlayerStateRepository
