+++ src/main/kotlin/dev/mizarc/waystonewarps/application/actions/discovery/GetPlayerWarpAccess.kt
@@ -1,12 +1,13 @@
-package dev.mizarc.waystonewarps.application.actions.warp
+package dev.mizarc.waystonewarps.application.actions.discovery
 
 import dev.mizarc.waystonewarps.domain.discoveries.DiscoveryRepository
 import dev.mizarc.waystonewarps.domain.warps.Warp
 import dev.mizarc.waystonewarps.domain.warps.WarpRepository
-import java.util.*
+import java.util.UUID
 
 class GetPlayerWarpAccess(private val discoveryRepository: DiscoveryRepository,
-                          private val warpRepository: WarpRepository) {
+                          private val warpRepository: WarpRepository
+) {
     fun execute(playerId: UUID): List<Warp> {
         val discoveries = discoveryRepository.getByPlayer(playerId)
         val warps = mutableListOf<Warp>()
