+++ src/main/kotlin/dev/mizarc/waystonewarps/application/actions/discovery/GetWarpPlayerAccess.kt
@@ -1,7 +1,7 @@
-package dev.mizarc.waystonewarps.application.actions.warp
+package dev.mizarc.waystonewarps.application.actions.discovery
 
 import dev.mizarc.waystonewarps.domain.discoveries.DiscoveryRepository
-import java.util.*
+import java.util.UUID
 
 class GetWarpPlayerAccess(private val discoveryRepository: DiscoveryRepository) {
     fun execute(warpId: UUID): List<UUID> {
