+++ src/main/kotlin/dev/mizarc/waystonewarps/application/actions/management/UpdateWarpIcon.kt
@@ -1,12 +1,12 @@
-package dev.mizarc.waystonewarps.application.actions.warp
+package dev.mizarc.waystonewarps.application.actions.management
 
 import dev.mizarc.waystonewarps.domain.warps.WarpRepository
 import java.util.*
 
-class UpdateWarpName(private val warpRepository: WarpRepository) {
-    fun execute(warpId: UUID, name: String): Result<Unit> {
+class UpdateWarpIcon(private val warpRepository: WarpRepository) {
+    fun execute(warpId: UUID, materialName: String): Result<Unit> {
         val warp = warpRepository.getById(warpId) ?: return Result.failure(Exception("Warp not found"))
-        warp.name = name
+        warp.icon = materialName
         warpRepository.update(warp)
         return Result.success(Unit)
     }
