+++ src/main/kotlin/dev/mizarc/waystonewarps/interaction/listeners/WaystoneDestructionListener.kt
@@ -1,20 +1,16 @@
 package dev.mizarc.waystonewarps.interaction.listeners
 
 import com.destroystokyo.paper.event.block.BlockDestroyEvent
-import dev.mizarc.waystonewarps.application.actions.warp.BreakWarpBlock
-import dev.mizarc.waystonewarps.application.actions.warp.GetWarpAtPosition
+import dev.mizarc.waystonewarps.application.actions.world.BreakWarpBlock
+import dev.mizarc.waystonewarps.application.actions.world.GetWarpAtPosition
 import dev.mizarc.waystonewarps.application.results.BreakWarpResult
 import dev.mizarc.waystonewarps.domain.warps.Warp
-import dev.mizarc.waystonewarps.infrastructure.mappers.toLocation
 import dev.mizarc.waystonewarps.infrastructure.mappers.toPosition3D
 import net.kyori.adventure.text.Component
-import net.kyori.adventure.text.TextComponent
 import net.kyori.adventure.text.format.TextColor
-import org.bukkit.Bukkit
 import org.bukkit.Material
 import org.bukkit.NamespacedKey
 import org.bukkit.block.Block
-import org.bukkit.entity.Entity
 import org.bukkit.entity.Player
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
@@ -28,7 +24,6 @@ import org.bukkit.inventory.ItemStack
 import org.bukkit.persistence.PersistentDataType
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
-import org.w3c.dom.Text
 import java.util.*
 
 class WaystoneDestructionListener: Listener, KoinComponent {
